ai-data-agent/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openai
python-dotenv
pydantic
